# docs/swagger.yaml
basePath: /
definitions:
  model.Subscription:
    properties:
      end_date:
        type: string
      price:
        type: integer
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API for manage users' subscriptions
  title: Subscription Service API
  version: "1.0"
paths:
  /subscriptions:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Subscription'
            type: array
      summary: Get all subscriptions
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      parameters:
      - description: Subscription
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/model.Subscription'
      produces:
      - application/json
      responses:
        "201":
          description: Created
        "400":
          description: Bad Request
      summary: Create new subscription
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      parameters:
      - description: ID
        in: path
        name: id
        required: true
        type: integer
      responses:
        "204":
          description: No Content
      summary: Delete subscription by ID
      tags:
      - subscriptions
  /subscriptions/total:
    get:
      parameters:
      - description: User ID
        in: query
        name: user
        required: true
        type: string
      - description: Service name
        in: query
        name: service
        type: string
      - description: Start date (MM-YYYY)
        in: query
        name: start
        required: true
        type: string
      - description: End date (MM-YYYY)
        in: query
        name: end
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
      summary: Total of all subscriptions
      tags:
      - subscriptions
swagger: "2.0"